Added business logic for Reset Layout button and renamed interfaces to elements
